chore: move `AnalogClock` to the shared library
