+++ edge-apps/blueprint/ts/components/AnalogClock.vue
