+++ edge-apps/blueprint/ts/components/index.ts
@@ -1,4 +1,5 @@
 export { default as InfoCard } from './InfoCard.vue'
 export { default as PrimaryCard } from './PrimaryCard.vue'
 export { default as BrandLogoCard } from './BrandLogoCard.vue'
+export { default as AnalogClock } from './AnalogClock.vue'
 export * from './icons'
