+++ edge-apps/clock/src/components/index.ts
@@ -1,4 +1,3 @@
-export { default as AnalogClock } from './AnalogClock.vue'
-export { default as DigitalTime } from './DigitalTime.vue'
-export { default as DateDisplay } from './DateDisplay.vue'
-export { default as InfoCard } from './InfoCard.vue'
+export { default as DigitalTime } from '@/components/DigitalTime.vue'
+export { default as DateDisplay } from '@/components/DateDisplay.vue'
+export { default as InfoCard } from '@/components/InfoCard.vue'
